+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/angular
 
+## 1.1.0-beta.7
+
+### Patch Changes
+
+- Updated dependencies [5a4e732]
+  - ai@5.1.0-beta.7
+
+## 1.1.0-beta.6
+
+### Patch Changes
+
+- Updated dependencies [0c4822d]
+  - ai@5.1.0-beta.6
+  - @ai-sdk/provider-utils@3.1.0-beta.2
+
 ## 1.1.0-beta.5
 
 ### Patch Changes